refactor(nim-serving): reorganize into api/accounts/ and pages/ structure
Group API code by resource under api/accounts/ (k8s, utils, hooks,
constants with colocated tests) and page components under
pages/projectSettings/. This structure accommodates future resource
groups like api/deployments/ from PR #7433.
